+++ src/main/java/com/kocak/kalah/service/PlayService.java
@@ -4,6 +4,6 @@ import com.kocak.kalah.model.dto.outgoing.BoardHeaderResponseDto;
 
 public interface PlayService {
 
-    BoardHeaderResponseDto makeMove(long gameId, short pit);
+    BoardHeaderResponseDto makeMove(long gameId, int pit);
 
 }
